updates selectUpcomingTrainings
