+++ src/components/training/AlternativeTrainings.js
@@ -2,9 +2,18 @@ import React from 'react'
 
 import { Card } from '../elements'
 import { H3 } from '../text'
-import { UpcomingTrainings } from '../training/UpcomingTrainingSection'
+import { UpcomingTrainings } from './UpcomingTrainingSection'
+import { getUpcomingTrainingsByType } from './UpcomingTrainings'
 import { LinkButton } from '../buttons'
 import { Col, Row } from '../layout/Grid'
+
+import {
+  ADVANCED_REACT,
+  PART_TIME,
+  REACT_WORKSHOP,
+  REACT_FUNDAMENTALS,
+  ONE_DAY_WORKSHOP,
+} from 'src/config/data'
 
 const AlternativeTrainings = ({
   trainings,
@@ -23,5 +32,26 @@ const AlternativeTrainings = ({
       </Row>
     </Card>
   ) : null
+
+export const AlternativeBootcampTrainings = ({ trainings }) => {
+  const reactTrainings = getUpcomingTrainingsByType({
+    trainings,
+    types: [
+      REACT_FUNDAMENTALS,
+      REACT_WORKSHOP,
+      ADVANCED_REACT,
+      ONE_DAY_WORKSHOP,
+      PART_TIME,
+    ],
+    first: 3,
+  })
+
+  return (
+    <AlternativeTrainings
+      trainings={reactTrainings}
+      titleText="Alternatives to the React Bootcamp"
+    />
+  )
+}
 
 export default AlternativeTrainings
